Switch points to big integer

+++ database/migrations/2017_10_10_10000_create_point_transactions_table.php
@@ -11,8 +11,8 @@ class CreatePointTransactionsTable extends Migration
             $table->increments('id');
             $table->text('message');
             $table->morphs('pointable');
-            $table->double('amount');
-            $table->double('current');
+            $table->unsignedBigInteger('amount');
+            $table->unsignedBigInteger('current');
             $table->timestamps();
         });
     }
